vision_server/api/requirements.txt
flask==3.1.1
flask-cors==6.0.1
face_recognition==1.3.0
opencv-python==4.12.0.88
pillow==10.4.0
numpy==2.2.6